% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_read.R
\name{add_snps}
\alias{add_snps}
\title{Adds n snps to each sequence in seq_dat}
\usage{
add_snps(seq_dat, n)
}
\arguments{
\item{seq_dat}{The sequences in which to introduce the snps}

\item{n}{The number of snps to introduce in each sequence}
}
\description{
Adds n snps to each sequence in seq_dat
}

