% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/align_sequences.R
\name{align_sequences}
\alias{align_sequences}
\title{A wrapper for a number of alignment tools}
\usage{
align_sequences(seqs, technique = "muscle", params = list())
}
\arguments{
\item{seqs}{The sequences to align}

\item{technique}{Which alignment program to use}

\item{params}{A list of parameters to pass to the alignment program}
}
\description{
A wrapper for a number of alignment tools
}

