% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bin_by_name.R
\name{bin_by_name}
\alias{bin_by_name}
\title{Groups sequences together into bins based on their names}
\usage{
bin_by_name(seq_dat)
}
\description{
Groups sequences together into bins based on their names
}

