% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batch_utilities.R
\name{bin_file}
\alias{bin_file}
\title{Bins a given FASTA file and outputs each bin as a seperate file}
\usage{

  bin_file(file_name = "~/projects/MotifBinner/data/CAP177_2040_v1merged.fastq",
  add_uniq_id = TRUE, number_of_front_bases_to_discard = 28,
  prefix = "CCAGCTGGTTATGCGATTCTMARGTG", suffix = "CTGAGCGTGTGGCAAGGCCC",
  motif_length = 9, max.mismatch = 5, fixed = FALSE,
  write_files = FALSE)
}
\arguments{
\item{file_name}{The file name}

\item{add_uniq_id}{If True, an integer will be appended to
the end of each sequence's name so that all identical sequences in a bin
gets the same number and sequences who are not identical will get different
numbers.}

\item{number_of_front_bases_to_discard}{The number of bases to remove from
the front of the sequence. The first few bases are part of the primer
sequence and needs to be trimmed off since they do not contain any extra
information. This varies between sequencing approaches, so the parameter
should be set with knowledge of the process. The primer can contain
degeneracies, so its better to chop it off earlier rather than later.}

\item{prefix}{See ?extract_motifs}

\item{suffix}{See ?extract_motifs}

\item{motif_length}{See ?extract_motifs}

\item{max.mismatch}{See ?extract_motifs}

\item{fixed}{See ?extract_motifs}

\item{write_files}{If this is a directory, the bins will be written to that
folder as FASTA files.}
}
\description{
Bins a given FASTA file and outputs each bin as a seperate file
}

