% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classify_mislabelling.R
\name{classify_bin}
\alias{classify_bin}
\title{Classify a bin's sequences into 'src' and 'out'}
\usage{
classify_bin(bin, technique = "random", params = list(n = 0.2))
}
\arguments{
\item{bin}{The input bin as a single DNAStringSet.}

\item{technique}{A string selecting which technique to use for the
classification}

\item{params}{A list of parameters used by the specific
classification techniques}
}
\value{
A list with two elements 'src' and 'out'
}
\description{
Using a specified method, split a bin into a list of two DNAStringSets. One
containing the sequences that we believe represents reads of the molecule of
interest and the other representing reads that were from a different
molecule.
}

