% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classify_mislabelling.R
\name{classify_bin_infovar_balance}
\alias{classify_bin_infovar_balance}
\title{Finds the outliers in a bin by balancing the amount of information that is
lost with the variance reduction that is realized.}
\usage{
classify_bin_infovar_balance(bin, threshold)
}
\arguments{
\item{bin}{The input bin as a single DNAStringSet.}

\item{threshold}{Outlier sequences are removed from the bin until the
distanct / information ratio drops below this threshold.}
}
\description{
This approach removes outliers from a bin by finding the sequence with the
highest average distance to all other sequences in the bin. It is then
removed. The ratio of the reduction in the average distance between the
sequences and the reduction in the information available is then computed.
The process will continue until either all data has been removed or the
ratio drops below some threshold
}

