% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classify_mislabelling.R
\name{classify_bin_most_frequent}
\alias{classify_bin_most_frequent}
\title{Bins based on most frequent sequence}
\usage{
classify_bin_most_frequent(bin)
}
\description{
Bins based on most frequent sequence
}

