% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classify_mislabelling.R
\name{classify_bin_random}
\alias{classify_bin_random}
\title{Randomly classify a number of sequences to the 'out' group}
\usage{
classify_bin_random(bin, n)
}
\arguments{
\item{bin}{The input bin as a single DNAStringSet.}

\item{n}{The proportion of the bin to remove.}
}
\description{
A very poor strategy included as a baseline
}

