% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_motifs.R
\name{clean_matches}
\alias{clean_matches}
\title{Internal Function: Clean matches to the motifs}
\usage{
clean_matches(matches, pattern_length)
}
\arguments{
\item{matches}{The matches to the motifs which potentially contains multiple}

\item{pattern_length}{The minimum length that the match must be in order to
be kept}
}
\description{
Internal Function: Clean matches to the motifs
}

