% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_motifs.R
\name{clean_seq_data}
\alias{clean_seq_data}
\title{Internal Function: Clean sequence data for motif extraction}
\usage{
clean_seq_data(seq_data)
}
\arguments{
\item{seq_data}{The sequences whose motifs must be extracted}
}
\description{
Internal Function: Clean sequence data for motif extraction
}

