% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{create_scenario_data}
\alias{create_scenario_data}
\title{Compute the scenarios that are not already in the scenario cache}
\usage{
create_scenario_data(unique_scenarios, scenario_cache)
}
\arguments{
\item{unique_scenarios}{A list of the unique scenarios as produced by
list_unique_scenarios}

\item{scenario_cache}{The scenario cache list}
}
\description{
Compute the scenarios that are not already in the scenario cache
}

