% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batch_utilities.R
\name{dput_classified_binned_folder}
\alias{dput_classified_binned_folder}
\title{Reads in all files from a folder, assuming that they are classified binned
files}
\usage{
dput_classified_binned_folder(binned_folder)
}
\arguments{
\item{binned_folder}{The folder containing the binned classified files}
}
\description{
Reads in all files from a folder, assuming that they are classified binned
files
}

