% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/consensus_string.R
\name{easyConsensusString}
\alias{easyConsensusString}
\title{A custom consensus string constructor that allows the threshold to be
relaxed. It will assign a letter to a position as long as that letter occurs
in enough sequences to meet the threshold.}
\usage{
easyConsensusString(seqs, threshold = 0.501)
}
\arguments{
\item{seqs}{The aligned sequences to construct a consensus of}

\item{threshold}{The percentage of sequences that must have a letter in a
certain positon for it to be included in the consensus string}
}
\description{
A custom consensus string constructor that allows the threshold to be
relaxed. It will assign a letter to a position as long as that letter occurs
in enough sequences to meet the threshold.
}

