% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_motifs.R
\name{extract_motifs}
\alias{extract_motifs}
\title{Extracts motifs from a set of reads}
\usage{
extract_motifs(seq_data, prefix, suffix, motif_length, max.mismatch = 5,
  fixed = FALSE)
}
\description{
Extracts motifs from a set of reads
}

