% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_motifs.R
\name{extract_motifs_iterative}
\alias{extract_motifs_iterative}
\title{A wrapper for extract motifs parallel that will iteratively allow for more
mismatches in the prefix and suffix surrounding the pids.}
\usage{
extract_motifs_iterative(seq_data, prefix, suffix, motif_length,
  max.mismatch = 5, fixed = FALSE, ncpu = 6, job_size = NULL,
  max.suffix.chop = NULL, max.mismatch_start = 0, verbose = FALSE)
}
\arguments{
\item{seq_data}{The sequences whose motifs must be extracted}

\item{prefix}{The prefix that is used to identify the motif}

\item{suffix}{The suffix that is used to identify the motif}

\item{motif_length}{The length of the motif that forms the pid.}

\item{max.mismatch}{See ?vmatchPattern}

\item{fixed}{See ?vmatchPattern}

\item{ncpu}{The number of cores to use}

\item{job_size}{The number of sequences to group into a single job}

\item{max.suffix.chop}{The maximum amount whereby to shorten the suffix
defaults to 0.5*length(suffix)}

\item{max.mismatch_start}{What is the minimum number of mismatches to search
for?}
}
\description{
A wrapper for extract motifs parallel that will iteratively allow for more
mismatches in the prefix and suffix surrounding the pids.
}

