% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_motifs.R
\name{extract_motifs_par}
\alias{extract_motifs_par}
\title{A wrapper for extract motifs that will execute it in a parallel nature using
the specified number of cpus}
\usage{
extract_motifs_par(seq_data, prefix, suffix, motif_length, max.mismatch = 5,
  fixed = FALSE, ncpu = 6, job_size = NULL, verbose = FALSE)
}
\arguments{
\item{seq_data}{The sequences whose motifs must be extracted}

\item{prefix}{The prefix that is used to identify the motif}

\item{suffix}{The suffix that is used to identify the motif}

\item{motif_length}{The length of the motif that forms the pid.}

\item{max.mismatch}{See ?vmatchPattern}

\item{fixed}{See ?vmatchPattern}

\item{ncpu}{The number of cores to use}

\item{job_size}{The number of sequences to group into a single job}

\item{verbose}{If set to true a progress report will be given via file
creation in the sessions temp dir}
}
\description{
A wrapper for extract motifs that will execute it in a parallel nature using
the specified number of cpus
}

