% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/distances.R
\name{fast_stringDist}
\alias{fast_stringDist}
\title{Computes a distance matrix between sequences by using relative distances to
a few carefully selected sequences}
\usage{
fast_stringDist(x, anchors = 3)
}
\arguments{
\item{x}{The input sequences}

\item{anchors}{The number of sequences to use as anchors}
}
\description{
The first sequence in the input data is chosen and the pairwise distances
between all sequences and this sequence is computed. The sequence most distant
from the first sequence is then chosen as the next anchor. The process
continues by selecting the sequence furthest from all previous anchors as the
next anchor.
}
\details{
Once all the distances to the chosen number of anchors have been computed, a
distance matrix is constructed using the distances between the sequences and
the anchors. For very large numbers of sequences (>500) this realizes a big
speedup.
}

