% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batch_utilities.R
\name{file_to_consensus}
\alias{file_to_consensus}
\title{Bins and constructs consensus sequences for an entire fastq file}
\usage{

  file_to_consensus(file_name = "~/projects/MotifBinner/data/CAP177_2040_v1merged.fastq",
  ...)
}
\arguments{
\item{file_name}{The file name}
}
\description{
Running process_bin in parallel yields 3x improvement in execution speed on
my laptop (213 sec -> 64 sec)
}

