% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_read.R
\name{gen_and_contaminate_reads}
\alias{gen_and_contaminate_reads}
\title{Generate contaminated reads}
\usage{
gen_and_contaminate_reads(ref_seq, n_reads, error_rates, contam_seq, n_contam,
  seed = NULL, ...)
}
\arguments{
\item{ref_seq}{The true sequence from which to generate the 'src' reads}

\item{n_reads}{The number of reads to generate}

\item{error_rates}{The error profile as generated by gen_error_profile}

\item{contam_seq}{The sequence to generate the contaminated reads from}

\item{n_contam}{The number of contamination reads to include}

\item{seed}{The seed for the randomizer}
}
\description{
Basically calls gen_reads twice and mixes the results together
}

