% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_read.R
\name{gen_error_profile}
\alias{gen_error_profile}
\title{Generates a read error profile}
\usage{
gen_error_profile(technique = "uniform", params = list(read_len = 1000,
  A_err_rates = list(A = 0.997, C = 0.001, G = 0.001, T = 0.001), C_err_rates =
  list(C = 0.997, A = 0.001, G = 0.001, T = 0.001), G_err_rates = list(G =
  0.997, C = 0.001, A = 0.001, T = 0.001), T_err_rates = list(T = 0.997, C =
  0.001, G = 0.001, A = 0.001)))
}
\arguments{
\item{technique}{What technique should be used to generate the error
profile?}

\item{params}{A list of parameters to pass to the specific profile generator}
}
\description{
Generates a read error profile
}

