% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_read.R
\name{gen_error_profile_uniform}
\alias{gen_error_profile_uniform}
\title{Generates a read error profile by assigning a constant probability of an
error at all positions}
\usage{
gen_error_profile_uniform(read_len, A_err_rates, C_err_rates, G_err_rates,
  T_err_rates)
}
\arguments{
\item{read_len}{The length of the profile}

\item{A_err_rates}{The probabilities that an A will be read as each of the
four nucleotides. Must sum to 1.}

\item{C_err_rates}{The probabilities that a C will be read as each of the
four nucleotides. Must sum to 1.}

\item{G_err_rates}{The probabilities that a G will be read as each of the
four nucleotides. Must sum to 1.}

\item{T_err_rates}{The probabilities that a T will be read as each of the
four nucleotides. Must sum to 1.}
}
\description{
Generates a read error profile by assigning a constant probability of an
error at all positions
}

