% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_read.R
\name{gen_pid_search_scenario}
\alias{gen_pid_search_scenario}
\title{Simulate a scenario that can be used to test PID detection}
\usage{
gen_pid_search_scenario(seq_len, prefix_len, pid_len, suffix_len, prefix_snps,
  suffix_snps, suffix_chop)
}
\arguments{
\item{seq_len}{The length of the random sequence to put before the prefix}

\item{prefix_len}{The length of the prefix}

\item{pid_len}{The length of the pid}

\item{suffix_len}{The length of the suffix}

\item{prefix_snps}{The number of snps to introduce to the prefix}

\item{suffix_snps}{The number of snps to introduce to the suffix}

\item{suffix_chop}{The number of bases to chop from the end of the suffix}
}
\description{
Simulate a scenario that can be used to test PID detection
}

