% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_read.R
\name{gen_reads}
\alias{gen_reads}
\title{Generates a specified number of reads from a given sequence based on a given
error profile}
\usage{
gen_reads(ref_seq, n_reads, error_rates, name_prefix = "src")
}
\arguments{
\item{ref_seq}{The reference sequence to generate the reads from}

\item{n_reads}{The number of reads to generate}

\item{error_rates}{The error profile as generated by gen_error_profile}

\item{name_prefix}{The prefix to give to the sequence names}
}
\description{
Generates a specified number of reads from a given sequence based on a given
error profile
}

