% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_read.R
\name{gen_seq}
\alias{gen_seq}
\title{Simulates a random sequence of a given length}
\usage{
gen_seq(n)
}
\arguments{
\item{n}{The length of the desired sequence}
}
\description{
Simulates a random sequence of a given length
}

