% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_mislabel_test_data}
\alias{get_mislabel_test_data}
\title{Returns that dataset used for testing the mislabel detection}
\usage{
get_mislabel_test_data()
}
\description{
Returns that dataset used for testing the mislabel detection
}

