% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{list_unique_scenarios}
\alias{list_unique_scenarios}
\title{Given all the test cases list the unique scenarios}
\usage{
list_unique_scenarios(cases)
}
\arguments{
\item{cases}{list of lists of the cases}
}
\description{
Given all the test cases list the unique scenarios
}

