% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{load_or_initialize_cache}
\alias{load_or_initialize_cache}
\title{Given the location of a cache file, intialize the cache}
\usage{
load_or_initialize_cache(cache_file)
}
\arguments{
\item{cache_file}{The absolute path to the cache file}
}
\description{
Given the location of a cache file, intialize the cache
}

