% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/consensus_string.R
\name{mostConsensusString}
\alias{mostConsensusString}
\title{A custom consensus string constructor that just uses the letters that occurs
most frequently at each position to construct the consensus sequence}
\usage{
mostConsensusString(seqs)
}
\arguments{
\item{seqs}{The aligned sequences to construct a consensus of}
}
\description{
In the case that two or more letter occurs the same number of times, an
ambiguity character is constructed from them
}

