% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_read.R
\name{mutate_base}
\alias{mutate_base}
\title{Mutates a single base to another base}
\usage{
mutate_base(x)
}
\arguments{
\item{x}{The input letter}
}
\description{
Mutates a single base to another base
}

