% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_bin.R
\name{process_bin}
\alias{process_bin}
\title{Given a groups of sequences that were binned together, throw out outliers,
align them and construct a consensus sequence}
\usage{
process_bin(seqs, classification_technique = "infovar_balance",
  classification_params = list(threshold = 1))
}
\arguments{
\item{seqs}{The sequences that were binned together}

\item{classification_technique}{The technique to use to search for
mislabelled sequences}

\item{classification_params}{The parameters for the classification technique}
}
\description{
Given a groups of sequences that were binned together, throw out outliers,
align them and construct a consensus sequence
}

