% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batch_utilities.R
\name{process_file}
\alias{process_file}
\title{Processes a file into consensus bins}
\usage{
process_file(file_name, prefix = "CCAGCTGGTTATGCGATTCTMARGTG",
  suffix = "CTGAGCGTGTGGCAAGGCCC", motif_length = 9, max.mismatch = 5,
  fixed = FALSE, add_uniq_id = T)
}
\arguments{
\item{file_name}{The file name}
}
\description{
Processes a file into consensus bins
}

