% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim_read.R
\name{randomize_ambig}
\alias{randomize_ambig}
\title{Removes ambiguous letters from a sequence by replacing them with a randomly
selected letter they represent}
\usage{
randomize_ambig(seq_dat)
}
\arguments{
\item{seq_dat}{The sequence in which to remove the ambiguity characters}
}
\description{
Removes ambiguous letters from a sequence by replacing them with a randomly
selected letter they represent
}

