% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batch_utilities.R
\name{randomize_list}
\alias{randomize_list}
\title{Randomizes the order of the items of a list}
\usage{
randomize_list(x)
}
\arguments{
\item{x}{The list to randomize}
}
\description{
Randomizes the order of the items of a list
}

