% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batch_utilities.R
\name{read_classified_binned_file}
\alias{read_classified_binned_file}
\title{Reads a classified binned file and splits it into bins}
\usage{
read_classified_binned_file(file_name, prefixes = c("src", "out"))
}
\arguments{
\item{file_name}{Name of the file to process}

\item{prefixes}{The prefixes that indicate the different bins}
}
\description{
Reads a classified binned file and splits it into bins
}

