% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batch_utilities.R
\name{read_sequence_file}
\alias{read_sequence_file}
\title{Reads in a sequence file given a file name.}
\usage{
read_sequence_file(file_name)
}
\arguments{
\item{file_name}{The file name}
}
\description{
Detects if it is a FASTA or FASTQ file and reads it appropriately. Quality
data will be dropped.
}

