% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_motifs.R
\name{remove_motifs}
\alias{remove_motifs}
\title{Internal Function: Removes motifs from sequences given match information}
\usage{
remove_motifs(matches, prefix, suffix, matched_seq)
}
\arguments{
\item{prefix}{The prefix that is used to identify the motif}

\item{suffix}{The suffix that is used to identify the motif}

\item{matched_seq}{The sequences in which there were motifs.}

\item{matched}{A set of IRanges describing the matched motif locations}
}
\description{
Internal Function: Removes motifs from sequences given match information
}

