% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{run_all_tests}
\alias{run_all_tests}
\title{Runs all the tests}
\usage{
run_all_tests(cases, scenario_cache)
}
\arguments{
\item{cases}{All the test cases to run}

\item{scenario_cache}{The cache with the pre computed scenarios}
}
\description{
Runs all the tests
}

