% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{run_test}
\alias{run_test}
\title{Run a test case}
\usage{
run_test(scenario, seed, setup, scenario_cache)
}
\arguments{
\item{scenario}{The computed scenario to run the test on}

\item{seed}{The seed that was used to generate the scenario}

\item{setup}{A list containing parameters that will control how the bin is
processed}

\item{scenario_cache}{The cache with the pre computed scenarios}
}
\description{
Run a test case
}

