% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/batch_utilities.R
\name{save_bin_report}
\alias{save_bin_report}
\title{Given a report_dat data list, generate and save the reports/log files into
the output directory}
\usage{
save_bin_report(output, report_dat, prefix_for_names)
}
\arguments{
\item{output}{The directory where the binning results are to be saved}

\item{report_dat}{The binning results as produced by process_file}

\item{prefix_for_names}{Add this bit of text to the front of each
sequence in the resulting consensus sequences.}
}
\description{
Given a report_dat data list, generate and save the reports/log files into
the output directory
}

