% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classification_checker.R
\name{score_all_classifications}
\alias{score_all_classifications}
\title{Given a list of datasets and a classification strategy, apply the strategy
to all datasets and compute the metrics}
\usage{
score_all_classifications(test_bins, technique, params)
}
\arguments{
\item{test_bins}{The list of input test bins.}

\item{technique}{A string selecting which technique to use for the
classification}

\item{params}{A list of parameters used by the specific
classification techniques}
}
\description{
Given a list of datasets and a classification strategy, apply the strategy
to all datasets and compute the metrics
}

