% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classification_checker.R
\name{score_classification}
\alias{score_classification}
\title{Given a test dataset and a classification strategy, apply the strategy and compute metrics}
\usage{
score_classification(test_bin, technique, params)
}
\arguments{
\item{test_bin}{The input test bin as a list of two DNAStringSets.}

\item{technique}{A string selecting which technique to use for the
classification}

\item{params}{A list of parameters used by the specific
classification techniques}
}
\description{
Given a test dataset and a classification strategy, apply the strategy and compute metrics
}

