% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/consensus_checker.R
\name{score_consensus}
\alias{score_consensus}
\title{Given a test bin of sequences, benchmark how well the input consensus
sequence can be reconstructed}
\usage{
score_consensus(test_bin, classification_technique = "infovar_balance",
  classification_params = list(threshold = 1),
  alignment_technique = "muscle", alignment_params = list(),
  consensus_technique = "Biostrings::consensusString",
  consensus_params = list())
}
\arguments{
\item{test_bin}{The test bin to benchmark}

\item{classification_technique}{The technique to use to search for
mislabelled sequences}

\item{classification_params}{The parameters for the classification technique}

\item{alignment_technique}{The alignment technique to use}

\item{alignment_params}{The alignment parameters to use}

\item{consensus_technique}{The technique to use when generating the
consensus sequence}

\item{consensus_params}{The parameters for the consensus generator}
}
\description{
Given a test bin of sequences, benchmark how well the input consensus
sequence can be reconstructed
}

