% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/consensus_string.R
\name{twoStepConsensusString}
\alias{twoStepConsensusString}
\title{A custom consensus string constructor that uses a two step threshold method
to construct consensus sequences.}
\usage{
twoStepConsensusString(seqs, threshold_1 = 0.501, threshold_2 = 0.25)
}
\arguments{
\item{seqs}{The aligned sequences to construct a consensus of}

\item{threshold}{The percentage of sequences that must have a letter in a
certain positon for it to be included in the consensus string}
}
\description{
First, threshold_1 is applied to search for all positions where a letter
occur in that many or more sequences. (threshold_1 must be bigger than 0.5).
The letter that occurs at least that frequently will then get put in the
consensus sequence.
}
\details{
For those positions where there is not a letter that meets threshold_1,
threshold_2 will then be used to remove all letters that occur in less than
threshold_2 sequences. Using the letters that remain, an ambiguity character
will be constructed.
}

