% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classify_mislabelling.R
\name{classify_bin_infovar_balance}
\alias{classify_bin_infovar_balance}
\title{Finds the outliers in a bin by balancing the amount of information that is
lost with the variance reduction that is realized.}
\usage{
classify_bin_infovar_balance(bin, threshold, start_threshold = 0,
  max_sequences = 100)
}
\arguments{
\item{bin}{The input bin as a single DNAStringSet.}

\item{threshold}{Outlier sequences are removed from the bin until the
distanct / information ratio drops below this threshold.}

\item{start_threshold}{Only start the classification if maximum distance in
the sample is greater then this number of bases normalized to the length of
the sequences. A value of 0.01 means that the procedure will only start if
the maximum distance between two sequences is greater then 1 if the
sequences is exactly 100 bases long. TODO further normalize for number of
sequences}

\item{max_sequences}{The maximum number of sequences to use for the
computation of the distance matrix. If more sequences than this is present,
then randomly select this many sequences and run the classification
algorithm on them.}
}
\description{
This approach removes outliers from a bin by finding the sequence with the
highest average distance to all other sequences in the bin. It is then
removed. The ratio of the reduction in the average distance between the
sequences and the reduction in the information available (where the number
of DNA sequences in the bin is taken as the measure of the amount of
information) is then computed.  The process will continue until either all
data has been removed or the ratio drops below some threshold
}

