% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/classify_mislabelling.R
\name{classify_bin_most_frequent}
\alias{classify_bin_most_frequent}
\title{Classifies the sequences in a set by only choosing the most frequent
sequence.}
\usage{
classify_bin_most_frequent(bin)
}
\arguments{
\item{bin}{The input bin as a single DNAStringSet}
}
\description{
Given a set of sequences, pick the most frequently occurring sequence as the
only sequence that belongs to that bin and label all other sequences as
outliers.
}
\details{
If there is more than one sequence that occurs the maximum number of times,
it will return all the sequences that occur that number of times.
}

