% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/consensus_string.R
\name{construct_consensus}
\alias{construct_consensus}
\title{Constructs a consensus string using the specified technique}
\usage{
construct_consensus(seqs, technique = "Biostrings::consensusString",
  params = list())
}
\arguments{
\item{seqs}{The aligned sequences to construct a consensus of}

\item{technique}{The technique to use}

\item{params}{Extra parameters passed to the consensus string construction
technique}
}
\description{
This function is specifically designed to work in the situation of binning
sequences based on PIDS, so it handles sequence names in a specific way. It
assumes that the name of the first sequence is the name that needs to be
given to the consensus sequence. Furthermore, it will append and underscore
and the number of sequences that were used to construct the consensus to the
end of this name.
}

