% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/extract_motifs.R
\name{extract_motifs}
\alias{extract_motifs}
\title{Extracts motifs from a set of reads}
\usage{
extract_motifs(seq_data, prefix, suffix, motif_length, max.mismatch = 5,
  fixed = FALSE)
}
\arguments{
\item{seq_data}{The sequences whose motifs must be extracted}

\item{prefix}{The prefix that is used to identify the motif}

\item{suffix}{The suffix that is used to identify the motif}

\item{max.mismatch}{See ?vmatchPattern}

\item{fixed}{See ?vmatchPattern}
}
\description{
Extracts motifs from a set of reads
}

