% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_bin.R
\name{process_bin}
\alias{process_bin}
\title{Given a groups of sequences that were binned together, throw out outliers,
align them and construct a consensus sequence}
\usage{
process_bin(seqs, classification_technique = "infovar_balance",
  classification_params = list(threshold = 1, start_threshold = 0.02,
  max_sequences = 100), alignment_technique = "muscle",
  alignment_params = list(),
  consensus_technique = "Biostrings::consensusString",
  consensus_params = list(), remove_gaps = TRUE, strip_uids = FALSE)
}
\arguments{
\item{seqs}{The sequences that were binned together}

\item{classification_technique}{The technique to use to search for
mislabelled sequences}

\item{classification_params}{The parameters for the classification technique}

\item{alignment_technique}{The alignment technique to use}

\item{alignment_params}{The alignment parameters to use}

\item{consensus_technique}{The technique to use when generating the
consensus sequence}

\item{consensus_params}{The parameters for the consensus generator}

\item{remove_gaps}{If set to TRUE (the default, then gaps will be removed
from the consensus sequences)}

\item{strip_uids}{Remove the unique identifiers from the sequence. It is not
intelligent. The names will be split on '_' and the first and last pieces
will be kept.}
}
\description{
The output will be a list with a DNAString. The name of the DNAString in the
list will be the bin name and will contain information about how many
sequences were used to construct the consensus. See construct_consensus.
}

