% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/process_bin.R
\name{process_bin}
\alias{process_bin}
\title{Given a groups of sequences that were binned together, throw out oitliers,
align them and construct a consensus sequence}
\usage{
process_bin(seqs)
}
\arguments{
\item{seqs}{The sequences that were binned together}
}
\description{
Given a groups of sequences that were binned together, throw out oitliers,
align them and construct a consensus sequence
}

